+++ apps/journeys-admin/src/components/Editor/EditToolbar/Menu/JourneyLink/JourneyLink.stories.tsx
@@ -4,9 +4,9 @@ import { Meta, StoryObj } from '@storybook/react'
 
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 
-import { ApolloLoadingProvider } from '../../../../../test/ApolloLoadingProvider'
-import { simpleComponentConfig } from '../../../../libs/storybook'
-import { publishedJourney } from '../../data'
+import { ApolloLoadingProvider } from '../../../../../../test/ApolloLoadingProvider'
+import { simpleComponentConfig } from '../../../../../libs/storybook'
+import { publishedJourney } from '../../../data'
 
 import { JourneyLink } from './JourneyLink'
