+++ apps/journeys-admin/src/components/Editor/EditToolbar/Menu/JourneyLink/SlugDialog/SlugDialog.spec.tsx
@@ -4,7 +4,7 @@ import { SnackbarProvider } from 'notistack'
 
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 
-import { defaultJourney } from '../../../data'
+import { defaultJourney } from '../../../../data'
 
 import { JOURNEY_SLUG_UPDATE, SlugDialog } from '.'
