+++ apps/journeys-admin/src/components/Editor/EditToolbar/Strategy/Strategy.tsx
@@ -1,25 +1,63 @@
 import Button from '@mui/material/Button'
-import { ReactElement } from 'react'
+import { ReactElement, useEffect } from 'react'
 
 import BulbIcon from '@core/shared/ui/icons/Bulb'
+import Typography from '@mui/material/Typography'
+import { useTranslation } from 'react-i18next'
+import {
+  ActiveJourneyEditContent,
+  useEditor
+} from '@core/journeys/ui/EditorProvider'
+import { ActiveSlide } from '@core/journeys/ui/EditorProvider/EditorProvider'
+import { MenuItem } from '../../../MenuItem'
 
-export function Strategy(): ReactElement {
+interface StrategyProps {
+  variant: 'button' | 'list-item'
+  closeMenu?: () => void
+}
+
+export function Strategy({ variant, closeMenu }: StrategyProps): ReactElement {
+  const { t } = useTranslation('apps-journeys-admin')
+  const { dispatch } = useEditor()
+
+  function handleGoalsClick(): void {
+    dispatch({
+      type: 'SetJourneyEditContentAction',
+      component: ActiveJourneyEditContent.Action
+    })
+    dispatch({
+      type: 'SetActiveSlideAction',
+      activeSlide: ActiveSlide.Content
+    })
+    closeMenu?.()
+  }
+
   return (
-    <Button
-      variant="outlined"
-      color="secondary"
-      startIcon={<BulbIcon />}
-      sx={{
-        display: {
-          xs: 'none',
-          md: 'flex'
-        }
-      }}
-      onClick={handleGoalsClick}
-    >
-      <Typography variant="subtitle2" sx={{ py: 1 }}>
-        {t('Strategy')}
-      </Typography>
-    </Button>
+    <>
+      {variant === 'button' ? (
+        <Button
+          variant="outlined"
+          color="secondary"
+          startIcon={<BulbIcon />}
+          sx={{
+            display: {
+              xs: 'none',
+              md: 'flex'
+            }
+          }}
+          onClick={handleGoalsClick}
+        >
+          <Typography variant="subtitle2" sx={{ py: 1 }}>
+            {t('Strategy')}
+          </Typography>
+        </Button>
+      ) : (
+        <MenuItem
+          label="Strategy"
+          icon={<BulbIcon />}
+          onClick={handleGoalsClick}
+        />
+      )}
+    </>
   )
 }
